Kind of a working api with now playing check

+++ src/api/master_clock.rs
@@ -7,5 +7,5 @@ pub struct MasterClock {
     /// Current master deck if any
     pub deck: Option<Deck>,
     /// Current BPM
-    pub bpm: u16
+    pub bpm: f32,
 }
